+++ pkg/scheduler/framework/plugins/registry.go
@@ -32,7 +32,7 @@ import (
 	"k8s.io/kubernetes/pkg/scheduler/framework/plugins/noderesources"
 	"k8s.io/kubernetes/pkg/scheduler/framework/plugins/nodeunschedulable"
 	"k8s.io/kubernetes/pkg/scheduler/framework/plugins/nodevolumelimits"
-	"k8s.io/kubernetes/pkg/scheduler/framework/plugins/podgroupcount"
+	"k8s.io/kubernetes/pkg/scheduler/framework/plugins/podgrouppodscount"
 	"k8s.io/kubernetes/pkg/scheduler/framework/plugins/podtopologyspread"
 	"k8s.io/kubernetes/pkg/scheduler/framework/plugins/queuesort"
 	"k8s.io/kubernetes/pkg/scheduler/framework/plugins/schedulinggates"
@@ -72,7 +72,7 @@ func NewInTreeRegistry() runtime.Registry {
 		schedulinggates.Name:                 runtime.FactoryAdapter(fts, schedulinggates.New),
 		gangscheduling.Name:                  runtime.FactoryAdapter(fts, gangscheduling.New),
 		topologyaware.Name:                   runtime.FactoryAdapter(fts, topologyaware.New),
-		podgroupcount.Name:                   runtime.FactoryAdapter(fts, podgroupcount.New),
+		podgrouppodscount.Name:               runtime.FactoryAdapter(fts, podgrouppodscount.New),
 	}
 
 	return registry
